Moved TextUtils::removeFormattingCodes to NarrationUtils.
